+++ src/app/material-timepicker/components/timepicker-field/timepicker-period-selector/ngx-timepicker-period-selector.component.ts
@@ -1,6 +1,8 @@
-import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
+import { ChangeDetectionStrategy, Component, EventEmitter, Inject, Input, Output } from '@angular/core';
 import { TimePeriod } from '../../../models/time-period.enum';
 import { animate, style, transition, trigger } from '@angular/animations';
+import { TIME_LOCALE } from '../../../tokens/time-locale.token';
+import { Info } from 'luxon';
 
 @Component({
     selector: 'ngx-timepicker-period-selector',
@@ -22,13 +24,24 @@ import { animate, style, transition, trigger } from '@angular/animations';
 
 export class NgxTimepickerPeriodSelectorComponent {
 
-    @Input() selectedPeriod: TimePeriod;
     @Input() isOpened: boolean;
     @Input() disabled: boolean;
+    @Input()
+    set selectedPeriod(period: TimePeriod) {
+        if (period) {
+            const periods = [TimePeriod.AM, TimePeriod.PM];
+            this.localizedPeriod = this.meridiems[periods.indexOf(period)];
+        }
+    }
 
     @Output() periodSelected = new EventEmitter<TimePeriod>();
 
     period = TimePeriod;
+    meridiems = Info.meridiems({locale: this.locale});
+    localizedPeriod: string;
+
+    constructor(@Inject(TIME_LOCALE) private locale: string) {
+    }
 
     open(): void {
         if (!this.disabled) {