+++ src/app/material-timepicker/directives/ngx-timepicker.directive.ts
@@ -1,9 +1,10 @@
-import { Directive, ElementRef, forwardRef, HostListener, Input, OnChanges, OnDestroy, SimpleChanges } from '@angular/core';
+import { Directive, ElementRef, forwardRef, HostListener, Inject, Input, OnChanges, OnDestroy, SimpleChanges } from '@angular/core';
 import { NgxMaterialTimepickerComponent } from '../ngx-material-timepicker.component';
 import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
 import { Subscription } from 'rxjs';
 import { TimeAdapter } from '../services/time-adapter';
 import { DateTime } from 'luxon';
+import { TIME_LOCALE } from '../tokens/time-locale.token';
 
 const VALUE_ACCESSOR = {
     provide: NG_VALUE_ACCESSOR,
@@ -37,7 +38,7 @@ export class TimepickerDirective implements ControlValueAccessor, OnDestroy, OnC
     @Input()
     set min(value: string | DateTime) {
         if (typeof value === 'string') {
-            this._min = TimeAdapter.convertTimeToDateTime(value, this._format);
+            this._min = TimeAdapter.parseTime(value, {locale: this.locale, format: this.format});
             return;
         }
         this._min = value;
@@ -52,7 +53,7 @@ export class TimepickerDirective implements ControlValueAccessor, OnDestroy, OnC
     @Input()
     set max(value: string | DateTime) {
         if (typeof value === 'string') {
-            this._max = TimeAdapter.convertTimeToDateTime(value, this._format);
+            this._max = TimeAdapter.parseTime(value, {locale: this.locale, format: this.format});
             return;
         }
         this._max = value;
@@ -78,7 +79,7 @@ export class TimepickerDirective implements ControlValueAccessor, OnDestroy, OnC
             this.updateInputValue();
             return;
         }
-        const time = TimeAdapter.formatTime(value, this._format);
+        const time = TimeAdapter.formatTime(value, {locale: this.locale, format: this.format});
         const isAvailable = TimeAdapter.isTimeAvailable(
             time,
             <DateTime>this._min,
@@ -97,7 +98,10 @@ export class TimepickerDirective implements ControlValueAccessor, OnDestroy, OnC
     }
 
     get value(): string {
-        return this._value;
+        if (!this._value) {
+            return '';
+        }
+        return TimeAdapter.toLocaleTimeString(this._value, {format: this.format, locale: this.locale});
     }
 
     private _value = '';
@@ -113,11 +117,14 @@ export class TimepickerDirective implements ControlValueAccessor, OnDestroy, OnC
     private onChange: (value: any) => void = () => {
     }
 
-    constructor(private elementRef: ElementRef) {
+    constructor(private elementRef: ElementRef,
+                @Inject(TIME_LOCALE) private locale: string) {
     }
 
     private set defaultTime(time: string) {
-        this._timepicker.setDefaultTime(time);
+        const formattedTime = TimeAdapter.formatTime(time, {locale: this.locale, format: this.format});
+
+        this._timepicker.setDefaultTime(formattedTime);
     }
 
     onInput(value: string) {
@@ -166,7 +173,7 @@ export class TimepickerDirective implements ControlValueAccessor, OnDestroy, OnC
             this._timepicker.registerInput(this);
             this.timepickerSubscriptions.push(this._timepicker.timeSet.subscribe((time: string) => {
                 this.value = time;
-                this.onChange(this._value);
+                this.onChange(this.value);
                 this.onTouched();
             }));
             this.timepickerSubscriptions.push(