+++ src/app/material-timepicker/models/time-options.interface.ts
@@ -1,0 +1,5 @@
+import { LocaleOptions } from 'luxon';
+
+export interface TimeOptions extends LocaleOptions {
+    format?: number;
+}