+++ src/app/material-timepicker/ngx-material-timepicker.module.ts
@@ -1,4 +1,4 @@
-import { NgModule } from '@angular/core';
+import { ModuleWithProviders, NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { NgxMaterialTimepickerComponent } from './ngx-material-timepicker.component';
 import { FormsModule } from '@angular/forms';
@@ -31,6 +31,12 @@ import {
 import {
     NgxTimepickerPeriodSelectorComponent
 } from './components/timepicker-field/timepicker-period-selector/ngx-timepicker-period-selector.component';
+import { TimeLocalizerPipe } from './pipes/time-localizer.pipe';
+import { TIME_LOCALE } from './tokens/time-locale.token';
+import { TimeAdapter } from './services/time-adapter';
+import { TimeParserPipe } from './pipes/time-parser.pipe';
+import { ActiveHourPipe } from './pipes/active-hour.pipe';
+import { ActiveMinutePipe } from './pipes/active-minute.pipe';
 
 
 @NgModule({
@@ -67,8 +73,24 @@ import {
         NgxMaterialTimepickerThemeDirective,
         NgxTimepickerFieldComponent,
         NgxTimepickerTimeControlComponent,
-        NgxTimepickerPeriodSelectorComponent
+        NgxTimepickerPeriodSelectorComponent,
+        TimeLocalizerPipe,
+        TimeParserPipe,
+        ActiveHourPipe,
+        ActiveMinutePipe
+    ],
+    providers: [
+        {provide: TIME_LOCALE, useValue: TimeAdapter.DEFAULT_LOCALE}
     ]
 })
 export class NgxMaterialTimepickerModule {
+
+    static setLocale(locale: string): ModuleWithProviders {
+        return {
+            ngModule: NgxMaterialTimepickerModule,
+            providers: [
+                {provide: TIME_LOCALE, useValue: locale}
+            ]
+        };
+    }
 }